feat: amélioration du header mobile avec menu admin et z-index fixe
- Ajout d'un menu mobile complet avec tous les liens dashboard
- Ajout du lien admin dans le menu mobile avec vérification des permissions
- Header fixe avec z-index élevé (z-[100] et z-[101] pour le menu mobile)
- Refactorisation du Header en composant client avec wrapper server pour les translations
- Amélioration de l'UX mobile avec icônes et meilleure organisation
- Résolution des types TypeScript pour éliminer les warnings
